+++ sdk/metric/pipeline_registry_test.go
@@ -149,6 +149,7 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 		{Name: "foo", Kind: InstrumentKindCounter},
 		{Name: "foo", Kind: InstrumentKindUpDownCounter},
 		{Name: "foo", Kind: InstrumentKindHistogram},
+		{Name: "foo", Kind: InstrumentKindGauge},
 		{Name: "foo", Kind: InstrumentKindObservableCounter},
 		{Name: "foo", Kind: InstrumentKindObservableUpDownCounter},
 		{Name: "foo", Kind: InstrumentKindObservableGauge},
@@ -185,6 +186,12 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 			validate: assertHist[N](metricdata.DeltaTemporality),
 		},
 		{
+			name:     "Default/Delta/Gauge",
+			reader:   NewManualReader(WithTemporalitySelector(deltaTemporalitySelector)),
+			inst:     instruments[InstrumentKindGauge],
+			validate: assertLastValue[N],
+		},
+		{
 			name:     "Default/Delta/PrecomputedSum/Monotonic",
 			reader:   NewManualReader(WithTemporalitySelector(deltaTemporalitySelector)),
 			inst:     instruments[InstrumentKindObservableCounter],
@@ -219,6 +226,12 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 			reader:   NewManualReader(),
 			inst:     instruments[InstrumentKindHistogram],
 			validate: assertHist[N](metricdata.CumulativeTemporality),
+		},
+		{
+			name:     "Default/Cumulative/Gauge",
+			reader:   NewManualReader(),
+			inst:     instruments[InstrumentKindGauge],
+			validate: assertLastValue[N],
 		},
 		{
 			name:     "Default/Cumulative/PrecomputedSum/Monotonic",
@@ -306,6 +319,12 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 			reader:   NewManualReader(WithAggregationSelector(func(ik InstrumentKind) Aggregation { return AggregationDefault{} })),
 			inst:     instruments[InstrumentKindHistogram],
 			validate: assertHist[N](metricdata.CumulativeTemporality),
+		},
+		{
+			name:     "Reader/Default/Cumulative/Gauge",
+			reader:   NewManualReader(WithAggregationSelector(func(ik InstrumentKind) Aggregation { return AggregationDefault{} })),
+			inst:     instruments[InstrumentKindGauge],
+			validate: assertLastValue[N],
 		},
 		{
 			name:     "Reader/Default/Cumulative/PrecomputedSum/Monotonic",
@@ -697,6 +716,32 @@ func TestIsAggregatorCompatible(t *testing.T) {
 		{
 			name: "SyncHistogram and ExponentialHistogram",
 			kind: InstrumentKindHistogram,
+			agg:  AggregationBase2ExponentialHistogram{},
+		},
+		{
+			name: "SyncGauge and Drop",
+			kind: InstrumentKindGauge,
+			agg:  AggregationDrop{},
+		},
+		{
+			name: "SyncGauge and LastValue",
+			kind: InstrumentKindGauge,
+			agg:  AggregationLastValue{},
+		},
+		{
+			name: "SyncGauge and Sum",
+			kind: InstrumentKindGauge,
+			agg:  AggregationSum{},
+			want: errIncompatibleAggregation,
+		},
+		{
+			name: "SyncGauge and ExplicitBucketHistogram",
+			kind: InstrumentKindGauge,
+			agg:  AggregationExplicitBucketHistogram{},
+		},
+		{
+			name: "SyncGauge and ExponentialHistogram",
+			kind: InstrumentKindGauge,
 			agg:  AggregationBase2ExponentialHistogram{},
 		},
 		{